Copy more maven testing module to support maven 3 still with junit 5

+++ src/main/java/org/mybatis/maven/mvnmigrate/util/MavenOutputStream.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2010-2022 the original author or authors.
+ *    Copyright 2010-2024 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -55,7 +55,7 @@ public class MavenOutputStream extends OutputStream {
   @Override
   public void write(byte[] b, int off, int len) throws IOException {
     for (int i = off; i < len; i++) {
-      write((int) b[i]);
+      write(b[i]);
     }
   }
